commit 6

+++ src/controllers/user_controller.rs
@@ -1,9 +1,10 @@
-use actix_web::{delete, get, post, put, web, HttpResponse, Responder};
-use actix_web::web::{service, Data, Json};
+use actix_web::{delete, get, post, put, web, HttpRequest, HttpResponse};
+use actix_web::web::{Data, Json};
 use serde_json::json;
+use crate::midleware::permission::{Permission, Role};
 use crate::models::user_model::{CreateUser, UpdateUser, UserLoginDto};
 use crate::services::user_service::UserService;
-
+use crate::utill::jwt::{ has_permission_with_roles};
 
 //login user (POST /users)
 #[post("/login")]
@@ -26,7 +27,12 @@ pub async fn create_user_controller(service: Data<UserService>, create_user: Jso
 
 // Create a new user (GET/users)
 #[get("/get-all")]
-pub async fn get_all_users_controller(service: Data<UserService>) -> HttpResponse {
+pub async fn get_all_users_controller(service: Data<UserService>,req:HttpRequest) -> HttpResponse {
+
+    if !has_permission_with_roles(&req,&vec![Role::Admin],&Permission::Read) {
+        return HttpResponse::Forbidden().body("You do not have permission to get users.");
+    }
+
     match service.get_all_users_service().await {
         Ok(users) => HttpResponse::Ok().json(users),
         Err(e) => HttpResponse::InternalServerError().body(e.to_string())
@@ -35,7 +41,12 @@ pub async fn get_all_users_controller(service: Data<UserService>) -> HttpRespons
 
 //Update a user by ID (PUT /users/{id})
 #[put("/update/{id}")]
-pub async fn update_user_controller(service: Data<UserService>, update_user: Json<UpdateUser>, id: web::Path<String>) -> HttpResponse {
+pub async fn update_user_controller(service: Data<UserService>, update_user: Json<UpdateUser>, id: web::Path<String>,req:HttpRequest) -> HttpResponse {
+
+    if !has_permission_with_roles(&req,&vec![Role::Admin],&Permission::Write){
+        return HttpResponse::Forbidden().body("You do not have permission to update users.");
+    }
+
     println!("user id : {}", &id);
     match service.update_user_service(update_user.into_inner(), id.to_string()).await {
         Ok(Some(user)) => HttpResponse::Created().json(user),
@@ -46,7 +57,11 @@ pub async fn update_user_controller(service: Data<UserService>, update_user: Jso
 
 // Delete a user by ID (DELETE /users/{id})
 #[delete("/delete/{id}")]
-pub async fn delete_user_controller(service: Data<UserService>, id: web::Path<String>) -> HttpResponse {
+pub async fn delete_user_controller(service: Data<UserService>, id: web::Path<String>,req:HttpRequest) -> HttpResponse {
+
+    if !has_permission_with_roles(&req,&vec![Role::Admin],&Permission::Delete) {
+        return HttpResponse::Forbidden().body("You do not have permission to delete users.");
+    }
     match service.delete_user_service(id.to_string()).await {
         Ok(_) => HttpResponse::NoContent().finish(),
         Err(e) => HttpResponse::InternalServerError().body(format!("Database error: {}", e))
